requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==7.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.4.2  # https://github.com/redis/redis-py
hiredis==2.1.1  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.4.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
uvicorn[standard]==0.20.0  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0  # https://github.com/adamchainz/django-cors-headers
djangorestframework-simplejwt
# DRF-spectacular for api documentation
drf-spectacular==0.25.1  # https://github.com/tfranzel/drf-spectacular
#New
djoser==2.1.0
#ML dependancies
pandas==1.5.3
joblib==1.2.0
scikit-learn==1.2.2
opencv-python==4.5.4.58
tensorflow==2.12.0
tensorflow_hub==0.13.0
fcm-django==1.0.14
firebase-admin==6.1.0
